+++ app/src/main/java/org/simple/clinic/summary/PatientSummaryUpdate.kt
@@ -54,13 +54,14 @@ class PatientSummaryUpdate : Update<PatientSummaryModel, PatientSummaryEvent, Pa
       LogTeleconsultClicked -> logTeleconsultClicked(model)
       is MedicalOfficersLoaded -> next(model.medicalOfficersLoaded(event.medicalOfficers))
       ChangeAssignedFacilityClicked -> dispatch(OpenSelectFacilitySheet)
-      is NewAssignedFacilitySelected -> dispatch(DispatchNewAssignedFacility(event.facility), RefreshNextAppointment)
+      is NewAssignedFacilitySelected -> dispatch(DispatchNewAssignedFacility(event.facility))
       is PatientRegistrationDataLoaded -> patientRegistrationDataLoaded(model, event)
       NextAppointmentActionClicked -> dispatch(ShowScheduleAppointmentSheet(
           model.patientUuid,
           NEXT_APPOINTMENT_ACTION_CLICK,
           model.currentFacility!!
       ))
+      AssignedFacilityChanged -> dispatch(RefreshNextAppointment)
     }
   }