+++ app/src/main/java/org/simple/clinic/summary/assignedfacility/AssignedFacilityUpdate.kt
@@ -2,6 +2,7 @@ package org.simple.clinic.summary.assignedfacility
 
 import com.spotify.mobius.Next
 import com.spotify.mobius.Update
+import org.simple.clinic.mobius.dispatch
 import org.simple.clinic.mobius.next
 import org.simple.clinic.util.toNullable
 
@@ -17,6 +18,7 @@ class AssignedFacilityUpdate : Update<AssignedFacilityModel, AssignedFacilityEve
           model.assignedFacilityUpdated(event.facility),
           ChangeAssignedFacility(model.patientUuid, event.facility.uuid)
       )
+      AssignedFacilityChanged -> dispatch(NotifyAssignedFacilityChanged)
     }
   }
 }