+++ app/src/main/java/org/simple/clinic/summary/assignedfacility/UiActions.kt
@@ -1,4 +1,5 @@
 package org.simple.clinic.summary.assignedfacility
 
 interface UiActions {
+  fun notifyAssignedFacilityChanged()
 }