+++ app/src/test/java/org/simple/clinic/summary/assignedfacility/AssignedFacilityEffectHandlerTest.kt
@@ -3,6 +3,7 @@ package org.simple.clinic.summary.assignedfacility
 import com.nhaarman.mockitokotlin2.doReturn
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.verify
+import com.nhaarman.mockitokotlin2.verifyNoMoreInteractions
 import com.nhaarman.mockitokotlin2.verifyZeroInteractions
 import com.nhaarman.mockitokotlin2.whenever
 import org.junit.After
@@ -76,7 +77,17 @@ class AssignedFacilityEffectHandlerTest {
         assignedFacilityId = updatedAssignedFacilityId
     )
 
-    effectHandlerTestCase.assertNoOutgoingEvents()
+    effectHandlerTestCase.assertOutgoingEvents(AssignedFacilityChanged)
     verifyZeroInteractions(uiActions)
+  }
+
+  @Test
+  fun `when notify assigned facility effect is changed, then notify the assigned facility changed`() {
+    // when
+    effectHandlerTestCase.dispatch(NotifyAssignedFacilityChanged)
+
+    // then
+    verify(uiActions).notifyAssignedFacilityChanged()
+    verifyNoMoreInteractions(uiActions)
   }
 }